scripts: h3: m3: update compliance test for MC
On Gen3 the format need to be propagated in the the MC pipeline before
streaming can start. Also running the v4l2-compliance as we did on Gen2
SoC will fail since it tests changing the formats on the video device
and check that streaming is still working. On Gen3 and MC this will fail
with -EPIPE since v4l2-compliance do not propagate the format change in
the pipeline before it attempts to star the stream.

Add a new helpers to run the v4l2-compliance checker without the
changing the format (remove the -f flag) of the video device and use the
format propagation helper to setup the pipeline before running the
compliance test.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>