h3: split mc setup tests between H3 ES1.x and ES2.0
The routing for these boards are different and can be treated as two
different boards. Split testing into two files.

Also rework how to figure out if to skip a active link in group testing.
This is needed as ES2.0 routes are not as deterministic as on ES1.x. It
also makes the code more readable.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>