vin-tests: handle source field format alternate
Since it became apparent that the VIN driver not yet support to output
alternate field formats to user-space we have two options how to handle
this. Either use the VIN hardware feature to interlace the two fields
into a complete frame or to only output the top or bottom field to
user-space.

This patch chooses to use the interlace function and to output a
complete frame to user-space. In order to do so the size and field of
the format for the VIN media node need to be adjusted if the source
format is alternate.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>