tools: edid: add local copy of edid helper
This a tool written by Laurent Pinchart and is copied from his repo at

  git://git.ideasonboard.com/xilinx/edid.git

Since this tool is not usually part of any distributions copy it to the
vin-tests repo to ease testing. Once v4l-utls learns how to work with
subdevicees this tool should be dropped for that solution.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>