vin-tests: set-edid: fix usage of mc_get_dev()
The set-edid utility uses the mc_get_dev() helper to find the ADV748x
HDMI sub-device. The recent improvement of mc_get_dev() changed behavior
of that helper to not search sub-device sysfs entires for matches which
is needed for this utility.

Also the string returned changed format from 'v4l-subdev42' to
/dev/v4l-subdev42', update the set-edid tool to handle this.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>