tools: vin_to_du: add Gen3 support
We can't deal with the pipeline configuration on Gen3 as it might
connect to the Salvator-X onboard ADV7482 CSI-2 transmitter or the
Salvator-X extension board for GMSL cameras. Same issue exists for V3M
Eagle, but there the GMSL pipeline is on the Eagle and ADVxxxx
components are on an expansion board.

One could look at the media pipeline and figure this out, but for now
lets just print a message informing the user that it needs to configure
the pipeline manually.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>