run-tests: only ask for user input at the end
The reason user input is useful is when running the tests in a git
bisect run. And when not it's not super annoying to have it there. It
should of course be made into a option to switch it on but do that
later. For now just move it to the end of the test run.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>