boards: add partial parallel input for V3M
The parallel input on the V3M is on the expansion board and requires out
tree DT patches and switches to be flipped on the Eagle board.

Further more the media bus format used in this patch is wrong, but I
can't find a format which is supported by the adv7604 driver and how the
ADV7612 is connected to the SoC. This media bus format is 16 bits wide
while only 12 bits are connected so colours are all wrong but at least
you see a image which is a approximation of the real thing.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>