gmsl-yavta: Set format on CSI-2 pad #0
The recent change to the R-Car CSI-2 receiver driver to
add support to the active state, disallowed setting a format
on the source pads, but relies instead on the format set on
the sink pad to be propagated to the source ones.

Adjust the gmsl-yavta.sh script to set format on the sink pad #0.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>