yavta-{cvbs,hdmi}: Fix how image size is fetched
The output from v4l2-ctl appears to have changed and there are now more
then one line containing the phrase Width. Update the string awk match
on to only capture the one we want.

Without this fix the script converting the captured frames fails as it
is feed bogus parameters.

Reported-by: Nguyen Hong Thuan <thuan.nguyen-hong@banvien.com.vn>
Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>